requirements.txt
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.7.5
mysql-connector==2.2.9
mysql-connector-python==8.4.0
nltk==3.8.1
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pillow==10.3.0
plotly==5.22.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.10
requests==2.31.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
tenacity==8.3.0
tqdm==4.66.4
tzdata==2024.1
urllib3==1.26.6
webencodings==0.5.1
zipp==3.18.2
